+++ Backend/TimeLoggerAPI/Models/TimeEntry.cs
@@ -12,7 +12,7 @@ public class TimeEntry
     public string? Description { get; set; }
     public bool IsOnSite { get; set; } = false;
     public decimal? TravelHours { get; set; }
-    public decimal? TravelKm { get; set; }
+    public int? TravelKm { get; set; }
     public DateTime CreatedDate { get; set; } = DateTime.UtcNow;
     public DateTime? ModifiedDate { get; set; }
